+++ test/runtests.jl
@@ -8,6 +8,8 @@ const TEST_MODE = get(ENV, "TEST_MODE", "default")
         include("unit/test_units.jl")
         include("unit/test_hamiltonian.jl")
         include("unit/test_energy.jl")
+        include("unit/test_binning.jl")
+        include("unit/test_observables.jl")
     end
     if TEST_MODE in ("default", "all", "aqua")
         include("aqua.jl")